Add transforms NormalizePercentile
Based on StarDist normalize

- Update readme, todo

+++ bioimageloader/utils.py
@@ -254,7 +254,7 @@ def expand_to_rgb(
         axis=-1,
     )
     if dtype is not None:
-        return stacked.astype(dtype)
+        return stacked.astype(dtype, copy=False)
     return stacked
